feat: Add a Set avoid inserting duplicated Orders
