+++ substreams.yaml
@@ -1,7 +1,7 @@
 specVersion: v0.1.0
 package:
   name: "decentraland_substreams"
-  version: v0.1.0
+  version: v0.0.3
 
 protobuf:
   files:
